Move checkHexadecimal into behavior file
